+++ Magic/CHANGELOG.md
@@ -13,6 +13,7 @@
  - Fix active costs for Air Scooter, also Levitate/Air Scooter and Boon will prevent mana regen while active.
  - Fix wand organizing and alphabetizing
  - The wand property "attributes" has been renamed to "item_attributes"
+ - The mob property "attributes" has been renamed to "entity_attributes"
 
 ## 7.3.1